feat(refactor): restructure the repos
