+++ batch_jobs/bin/__init__.py
@@ -1,0 +1,5 @@
+# import sys
+# import os
+# import bucket_manifest
+
+# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
