+++ batch_jobs/bin/run_object_metadata_job.py
@@ -1,0 +1,4 @@
+from bucket_manifest.object_metadata_job import run_job
+
+if __name__ == "__main__":
+    run_job()
