+++ batch_jobs/bucket_manifest/__init__.py
@@ -1,0 +1,2 @@
+from . import bucket_manifest_job
+from . import object_metadata_job
