+++ batch_jobs/bucket_manifest/bucket_manifest.Dockerfile
