+++ batch_jobs/bucket_manifest/object_metadata.Dockerfile
